requirements.txt
streamlit
pinecone-client
langchain
sentence-transformers
openai
requests
streamlit-chat
langchain-community
langchain-huggingface